Start splitting plugins.selection out of cli
