+++ letsencrypt/main.py
@@ -6,8 +6,6 @@ import os
 import sys
 import zope.component
 
-import letsencrypt
-
 from letsencrypt import account
 from letsencrypt import client
 from letsencrypt import cli
@@ -25,6 +23,7 @@ from letsencrypt import storage
 
 from letsencrypt.display import util as display_util, ops as display_ops
 from letsencrypt.plugins import disco as plugins_disco
+from letsencrypt.plugins.selection import choose_configurator_plugins
 
 import traceback
 import logging.handlers
@@ -404,7 +403,7 @@ def install(config, plugins):
     # this function ...
 
     try:
-        installer, _ = cli.choose_configurator_plugins(config, plugins, "install")
+        installer, _ = choose_configurator_plugins(config, plugins, "install")
     except errors.PluginSelectionError as e:
         return e.message
 
@@ -479,7 +478,7 @@ def run(config, plugins):  # pylint: disable=too-many-branches,too-many-locals
     # TODO: Make run as close to auth + install as possible
     # Possible difficulties: config.csr was hacked into auth
     try:
-        installer, authenticator = cli.choose_configurator_plugins(config, plugins, "run")
+        installer, authenticator = choose_configurator_plugins(config, plugins, "run")
     except errors.PluginSelectionError as e:
         return e.message
 
@@ -509,7 +508,7 @@ def obtain_cert(config, plugins, lineage=None):
     # pylint: disable=too-many-locals
     try:
         # installers are used in auth mode to determine domain names
-        installer, authenticator = cli.choose_configurator_plugins(config, plugins, "certonly")
+        installer, authenticator = choose_configurator_plugins(config, plugins, "certonly")
     except errors.PluginSelectionError as e:
         logger.info("Could not choose appropriate plugin: %s", e)
         raise
